refactor: change events schema (#13)
* Reformat the data structure for the events

Reformat so that we track the event count more easily, and we also track the year to use for leaderboards.

* Update the EventService

Update service to correctly update or create new entity

* Remove the updatedAt part of event entity

* Remove the get events method for now

We don't need this just yet and we have to do more work to fix with it around.

* Clean up the events entity

* Switch to use natural id

* Style changes

* Fix id used for log statement

* Rename year to event_year

year is a reserved keyword for the database, so we need to use something else.

* Fix integration tests